Move get_sample_data to a standalone function and script
other minor cleanup
